+++ ida-domain/src/main/kotlin/com/dgswiphak/ida/domain/secondary/model/value/Interview.kt
@@ -2,8 +2,8 @@ package com.dgswiphak.ida.domain.secondary.model.value
 
 data class Interview(
 
-    val studyCompetence: Int? = 0,
+    val studyCompetence: Double? = 0.0,
 
-    val computingCompetence: Int? = 0
+    val computingCompetence: Double? = 0.0
 
 )
